Fix: Add checks to ensure no configuration is overwritten

+++ install/installGitconfig.ps1
@@ -1,3 +1,11 @@
 # Install my git config file.
-New-Item -ItemType "directory" "$env:HOME\.config"
-Copy-Item -Recurse .\git "$env:HOME\.config\"
+if (!(Test-Path -Path "$env:HOME\.config")) {
+    New-Item -ItemType directory -Path "$env:HOME\.config"
+}
+
+if (!(Test-Path -Path "$env:HOME\.config\git")) {
+    Copy-Item -Recurse .\git "$env:HOME\.config\"
+}
+else {
+    Write-Host "It seems you already have an existing git configuration."
+}
